fix: bug when no selected wakeup
