+++ src/app/(login)/teacher/wakeup/TeacherWakeupContent.tsx
@@ -6,7 +6,6 @@ import { useRouter } from "next/navigation";
 import React from "react";
 import Swal from "sweetalert2";
 
-import { calcDateDiff } from "@/app/(login)/wakeup/utils";
 import { WakeupDB, WakeupGET, WakeupSelected } from "@/app/api/wakeup/utils";
 import { alert } from "@/utils/alert";
 import instance from "@/utils/instance";
@@ -26,7 +25,6 @@ const TeacherWakeupContent = ({ initailData }: {
   const [gender, setGender] = React.useState<"male" | "female">(initailData.gender);
   const [link, setLink] = React.useState<string>();
   const [selected, setSelected] = React.useState<WakeupSelected>(initailData.selected);
-  const [dateDiff, setDateDiff] = React.useState<string>("");
 
   const getWakeup = async () => {
     setLoading(true);
@@ -159,10 +157,6 @@ const TeacherWakeupContent = ({ initailData }: {
     setLoading(false);
   };
 
-  React.useEffect(() => {
-    setDateDiff(calcDateDiff(selected));
-  }, [selected.date]);
-
   const sunday = week.clone().day(0);
   const saturday = week.clone().day(6).add(1, "day");
 
@@ -200,11 +194,11 @@ const TeacherWakeupContent = ({ initailData }: {
           <tbody className="w-full border-y border-text/10 overflow-auto">
             <tr className="w-full">
               <th className="text-center px-4 whitespace-nowrap py-2 font-semibold w-full"
-                colSpan={2}>{gender === "male" ? "남" : "여"}학생 {dateDiff}의 기상송
+                colSpan={2}>{gender === "male" ? "남" : "여"}학생 {selected.dateDiff===undefined?"평행세계":selected.dateDiff}의 기상송
               </th>
             </tr>
             {
-              selected.id !== "" ? (
+              selected.id !== null ? (
                 <tr className="w-full border-y border-text/10">
                   <td></td>
                   <td
