+++ src/app/(login)/wakeup/list/ListContent.tsx
@@ -5,7 +5,6 @@ import moment from "moment";
 import { useRouter } from "next/navigation";
 import React from "react";
 
-import { calcDateDiff } from "@/app/(login)/wakeup/utils";
 import { WakeupDB, WakeupGET, WakeupSelected } from "@/app/api/wakeup/utils";
 import { alert } from "@/utils/alert";
 import instance from "@/utils/instance";
@@ -92,11 +91,11 @@ const ListContent = ({ initailData }: {
           <tbody className="w-full border-y border-text/10 overflow-auto">
             <tr className="w-full">
               <th className="text-centerpx-4 whitespace-nowrap py-2 font-semibold w-full"
-                colSpan={2}>{gender === "male" ? "남" : "여"}학생 {calcDateDiff(selected)}의 기상송
+                colSpan={2}>{gender === "male" ? "남" : "여"}학생 {selected.dateDiff===undefined?"평행세계":selected.dateDiff}의 기상송
               </th>
             </tr>
             {
-              selected.id !== "" ? (
+              selected.id !== null ? (
                 <tr className="w-full border-y border-text/10">
                   <td></td>
                   <td
