+++ src/app/(login)/wakeup/utils.ts
@@ -1,7 +1,3 @@
-import React from "react";
-
-import { WakeupSelected } from "@/app/api/wakeup/utils";
-
 export const menu = [
   {
     url: "/wakeup/list",
@@ -16,16 +12,3 @@ export const menu = [
     name: "내 신청",
   },
 ];
-
-export const calcDateDiff = (selected: WakeupSelected) => {
-  const today = new Date();
-  const day_selected = new Date(`${selected.date.substring(4,6)}/${selected.date.substring(6,8)}/${selected.date.substring(0,4)}`);
-  const diff = (Math.abs(today.getTime() - day_selected.getTime()) / (1000 * 60 * 60 * 24)).toFixed();
-
-  switch (diff) {
-  case "0": return "오늘";
-  case "1": return "어제";
-  case "2": return "그저께";
-  default:  return `${diff}일 전`;
-  }
-};
