+++ src/app/api/wakeup/utils.ts
@@ -15,6 +15,7 @@ export type WakeupSelected = {
   title: string;
   id: string;
   date: string;
+  dateDiff: string;
 }
 
 export type WakeupDB = WakeupData & {
